vinyl: abort all tasks on shutdown
Currently, on shutdown we ->complete tasks in output_queue and ->abort
tasks in the input_queue. As a result, we can't use event loop driven
things, like coeio, neither in ->complete nor in ->abort. This is OK for
->abort as there we only rollback index tree to the state it was before
the task was scheduled and are not going to do anything more than that.
On ->complete however we will need to write to metadata log, for which
we need coeio to avoid stalling the tx thread. One way to deal with that
is pass in_shutdown flag to ->complete as we used to not so long ago,
but it looks kinda ugly. Instead, let's just abort all tasks on
scheduler shutdown.