vinyl: prepare for multi-level compaction
Currently, we always compact all runs in a range. When multi-level
compaction is introduced we will select runs 1..M for compaction,
where M<=N, N is the total number of runs. The number of runs M depends
on compaction strategy. Keeping this in mind, this patch introduces
basic infrastructure for multi-level compaction. Now, the compact task
compacts runs starting from the newest one and ending with the run
returned by vy_range_get_compact_run(). For now the latter just returns
the oldest run in the range, so no functional changes are introduced by
this patch.